requirements.txt
httpx>=0.27
beautifulsoup4>=4.12
pillow>=10.0
google-genai>=1.0
python-dotenv>=1.0
jinja2>=3.1
